requirements.txt
Flask==3.0.3
Werkzeug==3.0.1
gunicorn==21.2.0

requests
python-whois

numpy>=2.0.0
scikit-learn==1.8.0
joblib==1.4.2

torch>=2.5.0
transformers>=5.0.0
einops>=0.7.0

soundfile>=0.12.0

opencv-python>=4.10.0.84
tqdm>=4.66.5
timm>=1.0.12

Pillow==10.3.0
pytesseract==0.3.10